signals_proyect/signals_proyect/requirements.txt
ccxt
pandas
pandas-ta
tabulate
colorama
PyYAML
numpy==1.26.4 
python-telegram-bot 